pyproject.toml
[tool.poetry]
name = "keyvee"
version = "0.1.0"
description = ""
authors = ["Kilari Teja <kilariteja9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"


[tool.poetry.group.dev.dependencies]
black = "^22.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
